Fix call to `parted mkpart`.
In PR #9932, a bug fix was made to ensure that GPT partitions don't
receive a default label of "primary". Unfortunately, that change relies
on a bug fix to `parted` that was made in v3.5, which is relatively
recent.

For reference:

- Ubuntu 22.04: 3.4
- Ubuntu 24.04: 3.6
- Azure Linux 2.0: 3.4
- Azure Linux 3.0: 3.6

This change provides a workaround by giving partitions a default name
of " ", if the build host has an older version of parted.